requirements.txt
pandas==2.2.3
numpy==2.2.5
scikit-learn==1.6.1
xgboost==3.0.0
joblib==1.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1

tqdm==4.67.1
colorama==0.4.6

matplotlib==3.10.1
seaborn==0.13.2

boto3==1.34.70